old broken logic in map_range for unmapping areas was failing in some cases - instead use unmap_pages; use stack closures in uses of traverse_entries so that we may use unmap_pages early in init before heap setup (use of transient wasn't appropriate anyway) (#951)
