+++ src/runtime/heap/debug_heap.c
@@ -1,4 +1,5 @@
 #include <runtime.h>
+#include <page.h>
 
 typedef struct dheap {
     struct heap h;
